Replace MQTT-C with everest::io:mqtt in framework MQTT abstraction
Use the mosquitto based everest::io::mqtt in the framework MQTT abstraction code

Make everest::io position independent

Also start mosquitto with IPv6 support in ci

Use disconnect event and subscribe/unsubscribe via add_action (this makes the operation threadsafe since it will run in the event loop)

Build io lib with bazel

Since it depends on mosquitto but the upstream recipe seems to have the wrong visibility set, patch this

Depend on io lib in framework to fix bazel build

Downgrade mosquitto version used in bazel build to the one used by cmake (2.0.22)

2.1.x seems to cause connectivity issues in our wrapper (needs investigating)

bazel: remove obsolete mqtt-c recipe

mqtt broker port is a 16 bit unsigned int, not just an int

Remove MQTT-C from framework dependencies.yaml

Remove mqtt-c from MODULE.bazel


fix: mosquitto using unix domain sockets


Disconnect MQTT connection before exiting manager

Otherwise the MQTT client can hang and manager does not exit at all


Removing verbose error message on repeat identical errors

+++ lib/everest/framework/everestrs/everestrs/src/lib.rs
@@ -177,7 +177,7 @@ mod ffi {
             prefix: &str,
             mqtt_broker_socket_path: &str,
             mqtt_broker_host: &str,
-            mqtt_broker_port: &u32,
+            mqtt_broker_port: &u16,
             mqtt_everest_prefix: &str,
             mqtt_external_prefix: &str,
         ) -> UniquePtr<Module>;
@@ -393,7 +393,7 @@ struct Args {
 
     /// MQTT broker port
     #[arg(long = "mqtt_broker_port")]
-    pub mqtt_broker_port: u32,
+    pub mqtt_broker_port: u16,
 
     /// MQTT EVerest prefix
     #[arg(long = "mqtt_everest_prefix")]
